// src/types.d.ts
export interface WorkOrderDashBoardSummary {
  summary: {
    totalWorkOrders: number,
    execution: WorkOrderSummary,
    survey: WorkOrderSummary,
  },
  weeklyData: WorkOrderWeeklyData[]
}

export interface WorkOrderSummary {
  total: number;
  pending: number;
  complete: number;
}

export interface WorkOrderWeeklyData {
  date: string,
  dayOfWeek: number,
  dayOfWeekThai: string,
  executionWorkOrders: number,
  fieldSurveys: number
}